// src/commons/HookGetSizeImage/enums.ts
export enum IMAGE_SIZE_ENUM {
  IS_360 = "IS_360",
  IS_481 = "IS_481",
  IS_721 = "IS_721",
  IS_1081 = "IS_1081",
  IS_1367 = "IS_1367",
  IS_1921 = "IS_1921",
  "IS_360@2X" = "IS_360@2X",
  "IS_481@2X" = "IS_481@2X",
  "IS_721@2X" = "IS_721@2X",
  "IS_1081@2X" = "IS_1081@2X",
  "IS_1367@2X" = "IS_1367@2X",
  "IS_1921@2X" = "IS_1921@2X",
}