// src/commons/HookGetSizeImage/interface.ts
export interface ImageNameWidthI {
  imageName: string;
}

export interface ImageMediaQueryI {
  is360: boolean;
  is481: boolean;
  is721: boolean;
  is1081: boolean;
  is1367: boolean;
  is1921: boolean;
  isMinDevicePixelRatio: boolean;
}

export interface ImageSizeI {
  IS_360: string;
  IS_481: string;
  IS_721: string;
  IS_1081: string;
  IS_1367: string;
  IS_1921: string;
  "IS_360@2X": string;
  "IS_481@2X": string;
  "IS_721@2X": string;
  "IS_1081@2X": string;
  "IS_1367@2X": string;
  "IS_1921@2X": string;
}